@@ -1,8 +1,8 @@
 module github.com/isometry/platform-health
 
-go 1.22.4
+go 1.23.0
 
-toolchain go1.22.5
+toolchain go1.23.4
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.30.3
