@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.35.2
+// 	protoc        v5.28.3
 // source: proto/detail_loop.proto
 
 package details
@@ -30,11 +30,9 @@ type Detail_Loop struct {
 
 func (x *Detail_Loop) Reset() {
 	*x = Detail_Loop{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_detail_loop_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_detail_loop_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Detail_Loop) String() string {
@@ -45,7 +43,7 @@ func (*Detail_Loop) ProtoMessage() {}
 
 func (x *Detail_Loop) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_detail_loop_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -96,7 +94,7 @@ func file_proto_detail_loop_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_detail_loop_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_proto_detail_loop_proto_goTypes = []interface{}{
+var file_proto_detail_loop_proto_goTypes = []any{
 	(*Detail_Loop)(nil), // 0: platform_health.detail.v1.Detail_Loop
 }
 var file_proto_detail_loop_proto_depIdxs = []int32{
@@ -111,20 +109,6 @@ func init() { file_proto_detail_loop_proto_init() }
 func file_proto_detail_loop_proto_init() {
 	if File_proto_detail_loop_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_proto_detail_loop_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Detail_Loop); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
