@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.35.2
+// 	protoc        v5.28.3
 // source: proto/detail_tls.proto
 
 package details
@@ -39,11 +39,9 @@ type Detail_TLS struct {
 
 func (x *Detail_TLS) Reset() {
 	*x = Detail_TLS{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_detail_tls_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_detail_tls_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Detail_TLS) String() string {
@@ -54,7 +52,7 @@ func (*Detail_TLS) ProtoMessage() {}
 
 func (x *Detail_TLS) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_detail_tls_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -183,7 +181,7 @@ func file_proto_detail_tls_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_detail_tls_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_proto_detail_tls_proto_goTypes = []interface{}{
+var file_proto_detail_tls_proto_goTypes = []any{
 	(*Detail_TLS)(nil),            // 0: platform_health.detail.v1.Detail_TLS
 	(*timestamppb.Timestamp)(nil), // 1: google.protobuf.Timestamp
 }
@@ -200,20 +198,6 @@ func init() { file_proto_detail_tls_proto_init() }
 func file_proto_detail_tls_proto_init() {
 	if File_proto_detail_tls_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_proto_detail_tls_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Detail_TLS); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
